Move Entity into include/, src/ files
