+++ world_render.hpp
@@ -1,89 +1,4 @@
-
-
-
-class Entity : public ElementID
-{
-private:
-  static Shader *flat_color_shader;
-  static bool shaders_initialized;
-
-  void initialize_shaders()
-  {
-    if (!shaders_initialized)
-    {
-      flat_color_shader = new Shader("data/shaders/flat_color_vertex.glsl",
-        "data/shaders/flat_color_fragment.glsl");
-      shaders_initialized = true;
-    }
-  }
-
-public:
-  placement3d place;
-  placement3d velocity;
-  Model3D *model;
-  ALLEGRO_BITMAP *texture;
-  int attached_script_id;
-
-  Entity(ElementID *manager, std::string id, Model3D *model=NULL, ALLEGRO_BITMAP *texture=NULL)
-    : ElementID(manager)
-    , place()
-    , velocity()
-    , model(model)
-    , texture(texture)
-    , attached_script_id(0)
-  {
-    initialize_shaders();
-
-    //set_id(id);
-    set(ENTITY_ID_ATTRIBUTE, id);
-    velocity.align = vec3d(0, 0, 0);
-    velocity.scale = vec3d(0, 0, 0);
-  }
-
-  void attach_script_id(int script_unique_id)
-  {
-    attached_script_id = script_unique_id;
-  }
-
-  int get_attached_script_id()
-  {
-    return attached_script_id;
-  }
-
-  bool is_script_attached()
-  {
-    return attached_script_id != 0;
-  }
-
-  void draw()
-  {
-    if (!model) return;
-    if (texture) model->set_texture(texture);
-
-    Shader::stop();
-
-    place.start_transform();
-    model->draw();
-    place.restore_transform();
-  }
-
-  void draw_flat_color(ALLEGRO_COLOR color)
-  {
-    if (!model) return;
-
-    flat_color_shader->use();
-    if (is_script_attached()) Shader::set_vec4("tint", color.r, color.g, color.b, color.a);
-    else Shader::set_vec4("tint", 0, 0, 0, 0);
-
-    place.start_transform();
-    model->draw();
-    place.restore_transform();
-  }
-};
-Shader *Entity::flat_color_shader = NULL;
-bool Entity::shaders_initialized = false;
-
-
+#include <entity.hpp>
 
 
 class WorldRenderScreen : public Screen
